feat: Add HTML highlighting to PHP strings
**Changes**
This commits add support for highlighting HTML inside of PHP strings.
For this to work it requires the user to put a comment containing HTML before the string.

**Context**
The reason we don't just do it for all strings, is because this messes up the default string highlighting.
We could however probably determine if a string contains HTML and then highlight the string.
For now this approach is fine though, so I went with this.